Moved more public API types to _user_mgt and aliased them
